refactor(dto): move ErrorResponseDto and PostalCodeDto from controller.global to dto.error
