+++ customers/src/main/java/io/github/joaoVitorLeal/marketsphere/customers/dto/brasilapi/PostalCodeDto.java
@@ -1,0 +1,4 @@
+package io.github.joaoVitorLeal.marketsphere.customers.dto.brasilapi;
+
+public record PostalCodeDto() {
+}
